[saveSettings] check for matching of code_unit and settings_unit
When airframe code_unit and settings unit are not matching, the parameter can't
be automatically saved and a warning is printed in the paparazzi center.

When they are matching, or if only one of the two is specified, it is
used as code value.

Values are displayed in settings alt_unit if any, and are stored in
airframe unit.