[ppm] check some rssi level to have a valid ppm frame
Default is always true
It is required on some receivers that are still produicing valid frames
after RC lost (futaba R617FS). It is still possible to get the lost
flag from an rssi reading or at least with a GPIO (usually possible when
the receiver has a led to indicate the RC lost state).