[rotorcraft] from Ewoud: change feedback gain scale in vertical control for better resolution The bit shifts are reduced by 5 (increasing the gain by factor 32), GUIDANCE_V_GAIN_SCALER (48) removed So to get the same gains as before, multiply them by 1.5
Maybe also adjust the GUIDANCE_V_MAX_SUM_ERR, which should really have a sensible default anyway!!!!
On the other hand the max sum error was not adjusted in any airframe file when the GAIN_SCALER was added by Bart 2012-04-03.
So now it's nearly back to the same level as we had then, which was probably good?