[rotorcraft] force MODE_STARTUP instead of KILL until ahrs is aligned
- Go to MODE_STARTUP (which defaults to AP_MODE_KILL) instead of AP_MODE_KILL as long as the ahrs is not aligned.
- Remove AUTOPILOT_DISABLE_AHRS_KILL from all example airframes (but kept the functionality for now).

MODE_STARTUP is set to AP_MODE_NAV for the autonomous (without RC) ARDrone2.
Without AUTOPILOT_DISABLE_AHRS_KILL it will force MODE_STARTUP until the ahrs is aligned,
but still not allow to turn on the motors (as that also checks for ahrs_is_aligned).

As far as I can see this should solve #964 as long as you don't have AUTOPILOT_DISABLE_AHRS_KILL defined.