[ocaml] [sim] Switched to dynamic linking on OS X.
An interesting lesson here is that ocamlc assumes that all .so files are
shared libraries. It does not do the same for .dylib files. This is why
we have to compile a dylib and save it away as a .so so that ocamlc is
able to do the right thing.