[gps] fix furuno configuration
The init function of the furuno GPS tried to add all settings messages to the transmit buffer in one go.
Since the uart buffer can't hold that many chars, only the first 4 or so config sentences were actually sent.

Added nmea_configure function that is in GpsEvent until configuration is finished.