[fixedwing] temporary fix for AHRS_TRIGGERED_ATTITUDE_LOOP
should be a temporary fix for runing attitude_loop on a new AHRS estimate for now.
This *should* work under the conditions that the ABI callback that binds to a message first,
is actually called last when the message is sent.
So when a new GYRO message is sent, it should first run the AHRS callbacks
(and update the state interface with the new estimate) and then set new_ins_attitude.

This should be properly implemented or removed before the stable release...