[rotorcraft] heading integration protection in RC-event
The PSI += .... was highly depending on the RC-frame rate.
With most RC this is not a problem but with variable message rates (e.g. rc_datalink dropping messages or joysticks sending too many frames) this line becomes dangerous.

So measure the time between to calls to read_rc_setpoint function instead of using a fixed RC_UPDATE_FREQ.
To prevent really weird things from happening, bound dt to 0.5s.

Don't partuculariliy like that this pulls in the sys_time dependency here,
but couldn't think of something better...

should close #452