[arch/linux] uart: thread only handles receiving data
uart_transmit should always be able to directly write the single byte to the serial device.
So there should be no need to use the tx queue.

Since the uart threads calls select on the fds to check when there is something to read,
it will not wake up if there is data in the tx buffer...

Using the tx buffer was previously done because the system timer was generating signals and interrupted the write.
Since this is no longer the case it should be fine like this.

Seems to work fine on the bebop and beagle bone black.