[modules] viewvideo: be nice to other threads
A simple attempt to slightly improve the scheduling, so that taking a shot doesn't block the rest so much.

This sets the computer vision thread nice level to 10, so that the main "thread" should get a little more.
It does in no way guarantee that something in the cv thread (like taking a pic) will not block the main functions like estimation, stabilization.

A better long term solution would probably be to run both the main program/thread and this module with the FIFO scheduler and give the cv thread a really low prio.

Not tested due to lack of hardware right now...