[rotorcraft] add NO_GPS_NEEDED_FOR_NAV
If a GPS is not used, and you want to use NAV mode, add to your airframe file:
<define name="NO_GPS_NEEDED_FOR_NAV" value="TRUE"/>
and make sure that you don't have a gps subsystem added.

This should close #1451 for now.
A better way to deal with this is probably still needed (e.g. if you want to be able to have a GPS connected, but not use it indoors).