[generators] always put configure options at beginning
put all makefile vars (specified via configure) before including the board makefile.

e.g. RADIO_CONTROL_PPM_PIN has moved to the board makefile, and this
```
<subsystem name="radio_control" type="ppm">
   <configure name="RADIO_CONTROL_PPM_PIN" value="UART1_RX"/>
</subsystem>
```
was not working anymore and instead it was necessary to move the configure line outside of the subsystem lines:
```
<subsystem name="radio_control" type="ppm"/>
<configure name="RADIO_CONTROL_PPM_PIN" value="UART1_RX"/>
```

This should fix that regression again.