[boards] support for OpenPilot Revolution board
IMU, mag, baro and telemetry tested.
Positive x-axis of the IMU is indicated by the arrows on the board, the side with LEDS and F4 is up (negative z-axis).
Still missing support for the modem and the external flash...

To get the axes correct also extended the imu_mpu600_hmc5883 driver to allow different axes/channel assignments.