[gps] call gps_x_init and event directly from module
and get rid of need for GPS registering.

PRIMARY_GPS (and SECONDARY_GPS) now needs to be UPPERCASE (the first part of the ABI id, e.g. GPS_UBX)
This is now only used/needed to resolve GpsId(PRIMARY_GPS) to e.g. GPS_UBX_ID

sim and nps are now failing, since the init and event functions are wrong for those targets.
Next step is to make gps_sim and gps_nps modules and remove those targets from the "normal" modules.
But maybe add the sim modules as dependencies so they can be auto-loaded?