[modules] add optical_flow_landing
I added an optical flow landing module for rotorcrafts.
It uses the size divergence from the optical flow algorithm to keep a constant divergence
(if set negative the drone will land, 0 = hover, and positive is going up).
Advantage with respect to a GPS landing is that it will take into account the landing surface,
even if it is higher than the ground level (so landing on a roof for instance).
The module does not use sonar for the vertical control, so it can be used by tiny drones that do not carry sonar.
The module uses a novel theory on seeing distances with just a single camera (no extra sensors) to detect when it should land.
So it can also be used by drones to estimate distance with a single camera.

closes #1611