The rules were not detecting Lisa/M V2.1 correctly.
The string in the Lisa/M V2.1 DFU bootloader is: Lisa/M V2.1 (Upgrade)
The wildcard now accepts the V2.1 as part of the string. Previous
versions of Lisa/M DFU bootloader did not have the hardware version in
them. This is needed so that the Lisa/M DFU is accessible from the
userspace in the plugdev group.