[modules] edgeflow noisemeasurement fix (#1820)
The more corners the algorithm tracked, the higher the noise measurement (est velocity error = tracked corners/ max_corners) becomes. However this is not wanted for the ins_hff, since there a higher estimated velocity error results in less trust in that measurementof the Kalman filter. I made a new equation, which actually gives a lower noise measurement with more corners.

In the future there needs to be a better, with optitrack groundtruth backed up, noise model, however this pull request will hopefully give an intermediate solution. I also added an optional median filter if the hff is not used or if more outliers needed to be filtered out.