[module] add shift_tracking module
This module is trying to track an axis based on a local position
estimate that can come from an UWB position system or a smart cam like
jevois for instance.
The control idea is to apply a shift around a reference trajectory that
should be close enough from the final desired path.
This was developped to help final approach for landing in a net. See
https://hal-enac.archives-ouvertes.fr/hal-01936955 for more information.
Also include a simple PID library (discrete or "continuous" form).