[ocaml] move the GCS-only related code out of xpprz lib
Most of the pprz.xlib is used by the legacy GCS only, using in
particular the deprecated canvans.
This PR is isolating the relevant part to the GCS folder, so it will be
easier in the future to avoid the compilation of this part by default.

Conditional compilation of the old GCS: test if glade is installed
to build supervision, logplotter and the (legacy) GCS.